# [libc++] Take advantage of trivial relocation in std::vector::erase