@ldionne 2024-11-14
Note that this is stacked on top of https://github.com/llvm/llvm-project/pull/116265.