@ldionne 2024-11-27
Rebased on top of main just to have a clean diff. I'll close this for now since I am actively working on coming up with trivial relocation facilities for libc++ and they differ a bit from what's in this patch.